+++ Common.Test/LoggerTest.cs
@@ -1,4 +1,5 @@
 using Common.Enum;
+using Common.FileHandlers;
 using Xunit;
 
 namespace Common.Test;
@@ -9,7 +10,7 @@ public class LoggerTest
     public void ShouldLogToFIle()
     {
         //Arrange
-        var logger = new Logger("Test");
+        var logger = new Logger(new FileWriteHandler(), "Test");
         //Act
         logger.Log("Test", "Test", LogType.INFO);
         var exists = System.IO.File.Exists($"./bin/pirateTest/logs/Test.log");