+++ Common/Logger.cs
@@ -1,4 +1,6 @@
 ﻿using Common.Enum;
+using Common.FileHandlers;
+using Common.FileHandlers.Interfaces;
 using Common.Interfaces;
 
 namespace Common;
@@ -6,32 +8,29 @@ namespace Common;
 [Serializable]
 public class Logger : ILogger
 {
-    private string logName { get; set; }
+    private string logFileName { get; set; }
     private string version { get; set; } = EnvironmentVariables.GetVariable("version");
     private string location { get; set; }
+    private readonly IFileWriteHandler _fileHandler;
 
-    public Logger(string Name = "placeholder")
+    public Logger(IFileWriteHandler FileHandler, string Name = "")
     {
-        logName = Name;
-        if (Name == "placeholder")
+        _fileHandler = FileHandler;
+
+        logFileName = Name;
+        if (Name == "")
         {
-            logName = $"{DateTime.Now.Day}.{DateTime.Now.Month}.{DateTime.Now.Year}.{DateTime.Now.Hour}.{DateTime.Now.Minute}.{DateTime.Now.Second}";
+            logFileName = $"{DateTime.Now.Day}.{DateTime.Now.Month}.{DateTime.Now.Year}.{DateTime.Now.Hour}.{DateTime.Now.Minute}.{DateTime.Now.Second}";
         }
-        location = $"./bin/pirate{version}/logs/{logName}.log";
 
-        bool exists = Directory.Exists($"./bin/pirate{version}/logs");
-        if (!exists)
-            Directory.CreateDirectory($"./bin/pirate{version}/logs");
-
-        var createdFile = File.Create(location);
-        createdFile.Close();
+        location = $"bin/pirate{version}/logs";
     }
 
     public void Log(string message, string orginFile, LogType logType)
     {
-        var file = File.AppendText(location);
         var time = DateTime.Now.ToString();
-        file.Write($"{time.Replace(" uur", "")}: {logType.ToString()}: {orginFile}.cs: {message}\n");
-        file.Close();
+        var text = ($"{time.Replace(" uur", "")}: {logType.ToString()}: {orginFile}.cs: {message}\n");
+
+        _fileHandler.AppendToFile(new FileWriteModel("log", ".log", location, text));
     }
 }