+++ Shell/Commands/CommandManager.cs
@@ -40,7 +40,7 @@ public class CommandManager : ICommandManager
                 Console.ForegroundColor = ConsoleColor.Red;
                 Console.WriteLine($"\n{exception}");
                 Console.ForegroundColor = ConsoleColor.White;
-                throw;
+                throw exception;
             }
         }
     }