+++ Shell/Commands/NewCommand.cs
@@ -4,8 +4,13 @@ namespace Shell.Commands;
 
 public class NewCommand : Command, ICommand, INewCommand
 {
-    public NewCommand(ILogger Logger) : base(Logger)
-    { }
+    private IFileWriteHandler _fileWriteHandler;
+    private IFileReadHandler _fileReadHandler;
+    public NewCommand(ILogger Logger, IFileWriteHandler FileWriteHandler, IFileReadHandler FileReadHandler) : base(Logger)
+    { 
+        _fileWriteHandler = FileWriteHandler;
+        _fileReadHandler = FileReadHandler;
+    }
     public override void Run(string[] arguments)
     {
         Logger.Log("Starting New Command", this.GetType().Name, LogType.INFO);
@@ -42,14 +47,10 @@ public class NewCommand : Command, ICommand, INewCommand
         switch (typeArgument)
         {
             case "gitignore":
-                var file = File.CreateText($"./.gitignore");
-                file.Write("[Bb]in/");
-                file.Close();
+                _fileWriteHandler.WriteToFile(new FileWriteModel("", ".gitignore",  "", "[Bb]in/"));
                 return;
             case "gitattributes":
-                file = File.CreateText("./.gitattributes");
-                file.Write("*.pirate linguist-language=Squirrel");
-                file.Close();
+                _fileWriteHandler.WriteToFile(new FileWriteModel("", ".gitattributes", "", "*.pirate linguist-language=Squirrel" ));
                 return;
             case "pirate":
                 var filename = "main";
@@ -58,14 +59,13 @@ public class NewCommand : Command, ICommand, INewCommand
                     filename = arguments[2];
                 }
                 catch (System.Exception) { }
-                var exists = File.Exists($"./{filename}.pirate");
-                if (exists)
+
+                if (_fileReadHandler.FileExists(filename, ".pirate", " "))
                 {
-                    Logger.Log($"Specified filename \"{filename}\" already exists", this.GetType().Name, LogType.WARNING);
                     Error($"Specified filename \"{filename}\" already exists");
                     return;
                 }
-                file = File.CreateText($"./{filename}.pirate");
+                _fileWriteHandler.WriteToFile(new FileWriteModel(filename, ".pirate", " ", ""));
                 Console.WriteLine($"\nCreated new .{typeArgument} file");
                 return;
         }