+++ CHANGELOG.md
@@ -53,6 +53,10 @@ Ordered by Release/Milestone, then pullrequest. Listing the changes made.
 - Completely rewrote lexer to remove the static methods and fields.
 - Renamed Lexer, Parser and Interpreter to PirateLexer, PirateParser and PirateInterpreter
 
+### ([#69](https://github.com/joerivanarkel/PirateLang/pull/69)) FileHandlers
+
+- Added a two filehandlers for writing and reading
+
 ## 0.1.2
 
 ### ([#32](https://github.com/joerivanarkel/PirateLang/pull/32)) Shell, CLI Improvements